Feature

+++ app/__init__.py
@@ -2,6 +2,7 @@
 # Importing objects from flask
 import os
 import jwt
+import datetime
 
 from flask import request, jsonify, render_template
 from flask_api import FlaskAPI
@@ -17,6 +18,7 @@ db = SQLAlchemy()
 def create_app(config_name):
     """ Wraps the creation of a new Flask object and returns it"""
     app = FlaskAPI(__name__, instance_relative_config=True)
+    app.url_map.strict_slashes = False
     # loads up config settings
     app.config.from_object(app_config[config_name])
     app.config.from_pyfile('config.py')
@@ -52,6 +54,8 @@ def create_app(config_name):
                 # Trying to decode the token found using the secret key
                 id = jwt.decode(token, os.getenv('SECRET'))['id']
                 current_user = User.query.filter_by(id=id).first()
+                if not current_user:
+                    return jsonify({'message': 'user not found!'}), 404
             except:
                 return jsonify({'message': 'Expired token! Please login'}), 401
 
@@ -60,7 +64,7 @@ def create_app(config_name):
 
         return decorated
 
-    @app.route('/auth/register', methods=['POST'])
+    @app.route('/auth/register/', methods=['POST'])
     def user_registration():
         """Method registers a user"""
         email = request.data['email']
@@ -96,15 +100,13 @@ def create_app(config_name):
         except:
             return jsonify({'error': 'An error occured!'}), 500
 
-    @app.route('/auth/login', methods=['POST'])
+    @app.route('/auth/login/', methods=['POST'])
     def user_login():
         email = request.data['email']
         password = request.data['password']
         found_user = User.query.filter_by(email=email).first()
         if not found_user:
-            resp = jsonify ({'error': 'User not found'})
-            resp.status_code = 401
-            return resp
+            return jsonify({'error': 'User not found'}), 401
 
         if not found_user.validate_password(password):
             resp = jsonify({'error': 'Wrong password!'})
@@ -114,12 +116,17 @@ def create_app(config_name):
         resp.status_code = 200
         return resp
 
-    @app.route('/auth/logout', methods=['POST'])
-    def logout():
+    @app.route('/auth/logout/', methods=['POST'])
+    @token_required
+    def logout(current_user):
         """Method logs out user"""
-        pass
+        token = jwt.encode({
+            'id': current_user.id,
+            'exp': datetime.datetime.utcnow()
+        }, os.getenv('SECRET'))
+        return jsonify({'token': token.decode('UTF-8')})
 
-    @app.route('/auth/reset_password', methods=['POST'])
+    @app.route('/auth/reset_password/', methods=['POST'])
     def reset_password():
         email = request.data['email']
         new_password = request.data['password']
@@ -129,21 +136,25 @@ def create_app(config_name):
             found_user.create_password(new_password)
             found_user.save()
             return jsonify({'message': 'Password has changed successfully'}), 200
-        return jsonify({'error': 'User not found!'}), 403
+        return jsonify({'error': 'User not found!'}), 404
 
-    @app.route('/auth/delete', methods=['POST'])
-    def delete_user():
-        email = str(request.data.get('email'))
-        found_user = User.query.filter_by(email=email).first()
+    @app.route('/auth/delete/', methods=['DELETE'])
+    @token_required
+    def delete_user(current_user):
+        password = str(request.data.get('password'))
+        found_user = User.query.filter_by(id=current_user.id).first()
         if not found_user:
             return jsonify({'error': 'User not found'}), 404
         else:
-            email = found_user.email
-            found_user.delete()
-            return jsonify({'message': 'User {} deleted'.format(email)}), 200
+            if password:
+                if found_user.validate_password(password):
+                    found_user.delete()
+                    return jsonify({'message': 'User is deleted'}), 200
+                return jsonify({'error': 'Please input correct password'}), 405
+            return jsonify({'error': 'Please input your password'}), 405
 
     # CRUD BU
-    @app.route('/bucketlist', methods=['POST', 'GET'])
+    @app.route('/bucketlist/', methods=['POST', 'GET'])
     @token_required
     def create_bucketlist(current_user):
 
@@ -171,7 +182,7 @@ def create_app(config_name):
             return jsonify({'error': 'Blank title. Please write your title'}), 401
         else:
             # GETs all bucketlists
-            url_endpoint = '/bucketlist'
+            url_endpoint = '/bucketlist/'
             search = request.args.get('q')
             page = int(request.args.get('page', default=1))
             # The page content limit should be 10
@@ -217,7 +228,7 @@ def create_app(config_name):
             bucket_dict['prev_page'] = prev_page
             return jsonify(bucket_dict), 200
 
-    @app.route('/bucketlist/<id>', methods=['GET'])
+    @app.route('/bucketlist/<id>/', methods=['GET'])
     @token_required
     def get_bucket(current_user, id):
         """Retrieves a buckelist using it's ID"""
@@ -234,7 +245,7 @@ def create_app(config_name):
             }
             return jsonify(resp), 200
 
-    @app.route('/bucketlist/<id>', methods=['DELETE'])
+    @app.route('/bucketlist/<id>/', methods=['DELETE'])
     @token_required
     def delete_bucket(current_user, id):
         """Deleting a specific bucketlist"""
@@ -242,13 +253,13 @@ def create_app(config_name):
         bucketlist = Bucketlist.query.filter_by(
             user_id=current_user.id, id=id).first()
         if not bucketlist:
-            return jsonify({'error': 'Bucketlist Not found'}), 403
+            return jsonify({'error': 'Bucketlist Not found'}), 404
         else:
             title = bucketlist.title
             bucketlist.delete()
             return jsonify({'message': 'Bucketlist {} deleted'.format(title)}), 200
 
-    @app.route('/bucketlist/<id>', methods=['PUT'])
+    @app.route('/bucketlist/<id>/', methods=['PUT'])
     @token_required
     def edit_bucket(current_user, id):
         """Edits a bucketlist"""
@@ -269,7 +280,7 @@ def create_app(config_name):
             return jsonify(resp), 200
 
     # CRUD BUCKET LIST ITEMS
-    @app.route('/bucketlist/<id>/item', methods=['POST'])
+    @app.route('/bucketlist/<id>/item/', methods=['POST'])
     @token_required
     def create_item(current_user, id):
         """Method creates an item"""
@@ -295,24 +306,9 @@ def create_app(config_name):
                     return jsonify({'error': 'Name already exists'}), 403
                 return jsonify({'error': 'Item name not given!'}), 401
         return jsonify({'error':'Bucketlist does not exist!'}), 403
-        # else:
-        #     # GETs all items in a list
-        #     items = Item.query.filter_by(bucket_id=id).all()
-        #     if not items:
-        #         return jsonify({'error': 'No bucket items found!'}), 403
 
-        #     results = []
-        #     for item in items:
-        #         obj = {
-        #             'id': item.id,
-        #             'name': item.name
-        #         }
-        #         results.append(obj)
-        #     resp = jsonify(results)
-        #     resp.status_code = 200
-        #     return resp
 
-    @app.route('/bucketlist/<id>/item/<item_id>', methods=['DELETE'])
+    @app.route('/bucketlist/<id>/item/<item_id>/', methods=['DELETE'])
     @token_required
     def delete_bucketlist_item(current_user, id, item_id):
         """"Deltes a bucketlist"""
@@ -323,7 +319,7 @@ def create_app(config_name):
             item.delete()
             return jsonify({'message': 'Bucketlist item deleted'})
 
-    @app.route('/bucketlist/<id>/item/<item_id>', methods=['PUT'])
+    @app.route('/bucketlist/<id>/item/<item_id>/', methods=['PUT'])
     @token_required
     def edit_bucketlist_item(current_user, id, item_id):
         """Edits a bucketlist"""
